+++ RELEASE_NOTES.md
@@ -2,9 +2,12 @@
 
 ### RELEASE 0.3.2-beta
 
-FIXES:
+ENHANCEMENTS:
 
+1. SPARTNReader will now store any 32-bit gnssTimeTags for each msgSubtype (GPS, GLO, GAL, etc.) from the incoming datastream for use as 'basedates' in the decryption of any encrypted messages with ambiguous 16-bit gnssTimetags (timeTagtype = 0). If no 32-bit gnssTimeTags are available for a given msgSubtype, the input argument 'basedate' will be used instead, but
+bear in mind this may only be valid for one particular msgSubtype (i.e. GNSS constellation) due to different UTC and leap second shifts.
 1. Update test cases.
+1. Other minor internal streamlining.
 
 ### RELEASE 0.3.1-beta